Move docker blockchain specific traits and structs in blockchain folder
